refacto(moderations): move lib internal code
