+++ package.json
@@ -48,6 +48,7 @@
     "@~/app.core": "workspace:*",
     "@~/app.middleware": "workspace:*",
     "@~/app.urls": "workspace:*",
+    "@~/moderations.lib": "workspace:*",
     "@~/moderations.repository": "workspace:*",
     "@~/moncomptepro.database": "workspace:*",
     "@~/zammad.api": "workspace:*",
