+++ packages/~/moderations/lib/src/Moderation_Type.ts
@@ -1,0 +1,7 @@
+//
+
+export type Moderation_Type =
+  | "ask_for_sponsorship"
+  | "big_organization_join"
+  | "non_verified_domain"
+  | "organization_join_block";
