+++ packages/~/moderations/lib/src/moderation_type.mapper.ts
@@ -1,12 +1,12 @@
 //
 
-import type { MCP_Moderation } from ":moncomptepro";
 import { match } from "ts-pattern";
+import type { Moderation_Type } from "./Moderation_Type";
 
 //
 
 export function moderation_type_to_emoji(type: string) {
-  return match(type as MCP_Moderation["type"])
+  return match(type as Moderation_Type)
     .with("ask_for_sponsorship", () => "🧑‍🤝‍🧑")
     .with("big_organization_join", () => "🏢")
     .with("non_verified_domain", () => "🔓")
@@ -15,7 +15,7 @@ export function moderation_type_to_emoji(type: string) {
 }
 
 export function moderation_type_to_title(type: string) {
-  return match(type as MCP_Moderation["type"])
+  return match(type as Moderation_Type)
     .with("ask_for_sponsorship", () => "Sponsorship")
     .with("big_organization_join", () => "Big Organisation")
     .with("non_verified_domain", () => "Non vérifié")
