+++ packages/~/moderations/lib/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "outDir": "./node_modules/.cache/tsc"
+  },
+  "extends": "@~/config.typescript/base.json"
+}
