+++ src/legacy/users/ModerationTable.tsx
@@ -1,8 +1,8 @@
 //
 
-import { moderation_type_to_emoji } from ":moderations/moderation_type_to_emoji";
 import { row } from ":ui/table";
 import { urls } from "@~/app.urls";
+import { moderation_type_to_emoji } from "@~/moderations.lib/moderation_type.mapper";
 import type { Moderation } from "@~/moncomptepro.database";
 import { createContext } from "hono/jsx";
 import { match } from "ts-pattern";
