+++ src/moderations/:id/Header.tsx
@@ -1,11 +1,11 @@
 //
 
-import { moderation_type_to_emoji } from ":moderations/moderation_type_to_emoji";
 import type { MCP_Moderation } from ":moncomptepro";
 import { button } from ":ui/button";
 import { callout } from ":ui/callout";
 import { LocalTime } from ":ui/time/LocalTime";
 import { urls } from "@~/app.urls";
+import { moderation_type_to_emoji } from "@~/moderations.lib/moderation_type.mapper";
 import type { Moderation } from "@~/moncomptepro.database";
 import { useContext } from "hono/jsx";
 import { match } from "ts-pattern";
