+++ src/moderations/:id/page.tsx
@@ -1,10 +1,10 @@
 //
 
-import { MODERATION_EVENTS } from ":moderations/event";
 import { button } from ":ui/button";
 import { hx_trigger_from_body } from "@~/app.core/htmx";
 import type { MonComptePro_Pg_Context } from "@~/app.middleware/moncomptepro_pg";
 import { urls } from "@~/app.urls";
+import { MODERATION_EVENTS } from "@~/moderations.lib/event";
 import { schema } from "@~/moncomptepro.database";
 import { and, eq } from "drizzle-orm";
 import { useRequestContext } from "hono/jsx-renderer";
